tests/local/tree-construction/remex01.dat

#data
<pre><!doctype html>
x</pre>
#errors
[0] missing doctype
[5] unexpected doctype
#document
| <html>
|   <head>
|   <body>
|     <pre>
|       "
x"

#data
<table><tr><td><table><tr id=1></table><tr id=2></tr></table>
#errors
[0] missing doctype
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <table>
|               <tbody>
|                 <tr>
|                   id="1"
|         <tr>
|           id="2"

#data
<!doctype html><table><tr><td><table><tr id=1></tbody><tbody id=2></table></td></tr><tr id=3></tr></table>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <table>
|               <tbody>
|                 <tr>
|                   id="1"
|               <tbody>
|                 id="2"
|         <tr>
|           id="3"

#data
<!doctype html><div><i><b>foo<b><p><b>foo</i></p>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <div>
|       <i>
|         <b>
|           "foo"
|           <b>
|       <b>
|         <b>
|           <p>
|             <i>
|               <b>
|                 "foo"

#data
<!doctype html><div id="&"></div><div id="&amp;"></div><div id="&amp;amp;"></div>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <div>
|       id="&"
|     <div>
|       id="&"
|     <div>
|       id="&amp;"

#data
<!doctype html><svg><g id="&amp;amp;"></g></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         id="&amp;"
